[inductor] split test_cpp_wrapper.py into cpu and cuda test files
[ghstack-poisoned]
